% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_CellChat_list}
\alias{example_CellChat_list}
\title{Example CellChat object list}
\format{
A named list of 5 CellChat objects
}
\usage{
example_CellChat_list
}
\description{
A list of 5 slimmed CellChat objects for use in examples and tests.
Each object contains a subset of cell types and signaling pathways.
}
\keyword{datasets}
