% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_DB}
\alias{example_DB}
\title{Example CellChat database}
\format{
A list containing the CellChat DB slot
}
\usage{
example_DB
}
\description{
The CellChat ligand-receptor database extracted from the example objects.
Store separately and re-attach to objects when running extract_cellchat_data.
}
\keyword{datasets}
